routing: update static routes with config package
Previously, the static routes were rendered with the pattern
`/\w+`. To ensure the user folder does not clash with the routing,
begin each static route with `/__/`.

And part of refactoring also means extracting these values as
resuable constants.

The HTML template is also injected with these values to allow
future configurability.