UI: load content only when page is first visited (#6)
This PR focus on reducing the number of calls to the refresh endpoint to save on load time. Previously, once you run the tool on multiple files, it will load all content for each tab as soon as it can. Instead, why not load the contents only if the user visits that page for the first time. That way for larger pages, no unnecessary calls to refresh will be triggered till needed.

You have to visit the page at least once to initiate a stream connection to receive the markdown contents. So if you make changes to a file that belongs to a unvisited tab, the logger will not register any changes since no connection is made yet. However, this is not an issue as when the user visits the page the first time, the latest contents of that fille will be loaded.

Reduce sleep time for the Watch() related tests to speed up the test runs.

vui-chee: load-content-on-visit